refactor(power): share task-size loading helpers
